Move channel state types to channel module
